requirements.txt
streamlit
pymongo
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
python-dotenv